% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferSubset.R
\name{bufferSubset}
\alias{bufferSubset}
\title{bufferSubset}
\usage{
bufferSubset(
  gtShp,
  remSensShp,
  crs = NULL,
  gtSampleID,
  remSensSampleID,
  remSensPatch,
  gpsAccuracy = NULL,
  gpsAccuracyField = NULL,
  proportionOverlap = 1
)
}
\arguments{
\item{gtShp}{A point shapefile of class \code{sf} with ground-truthing data,
should be in projected coordinate system (e.g., UTM, state plane).}

\item{remSensShp}{A polygon shapefile of class \code{sf} with remote sensing
classifications, should be in projected coordinate system (e.g., UTM, state
plane) and must match the \code{gtShp} coordinate system if \code{crs} is
\code{NULL}.}

\item{crs}{Optional. CRS to be used to reproject \code{gtShp} and
\code{remSensShp}.}

\item{gtSampleID}{The name of the \code{gtShp} field containing the
ground-truthing sampleID.}

\item{remSensSampleID}{The name of the \code{remSensShp} field containing the
sampleID for each classified patch.}

\item{remSensPatch}{The name of the \code{remSensShp} field containing the
remote sensing classifications.}

\item{gpsAccuracy}{Magnitude of the horizontal radius that encompasses 95\% of
the data, this is also referred to as distance root mean square error. If
\code{crs} is supplied, units must match the crs units, otherwise the units
must match the coordinate system units of the supplied shapefiles (e.g.,
meters for UTM coordinate systems).}

\item{gpsAccuracyField}{The name of the \code{gtShp} field containing the
ground-truthing GPS accuracy. If provided, gpsAccuracy will be ignored.}

\item{proportionOverlap}{The proportion of a buffered area that must fall
within a single patch type in order for the corresponding ground-truthing
data to be retained. Proportions must be greater than 0.5 to avoid potential
conflicts with multiple patch types being matched to individual
ground-truthing points.}
}
\value{
A shapefile of class \code{sf} with spatially joined ground-truthing
and remote sensing classifications, subset to only include data where
ground-truthing points with a buffered area, having a radius equal to
\code{gpsAccuracy}, have at least \code{proportionOverlap} proportion of
their buffered area within a single patch type.
}
\description{
Join and subset ground-truthing point shapefile and patch type polygon
shapefile to only include points where a buffered area (defined by
\code{gpsAccuracy}) is fully within a single patch type.
}
\details{
A radius equal to \code{gpsAccuracy} is drawn around each
ground-truthing point. Buffered  ground-truthing points are joined with the
polygons of the remote sensed patches and only ground-truthing points where
the area of the buffer within a single patch type is greater than the product
of the  \code{proportionOverlap} and the area of the buffer within polygons
of classified patches are retained, representing a data set where the degree
of confidence the two data types (ground-truthing points and remote sensing
polygons) overlap is high.
}
\examples{
data(substrateLayers)
bufferedSubset <- bufferSubset(gtShp=substrateLayers[['gtDom_sf']], 
remSensShp=substrateLayers[['sonarClass']], 
                               gtSampleID="samplID", remSensSampleID="id", 
                               remSensPatch="substrate", gpsAccuracy=5)
}
\seealso{
\link{classificationAccuracy_propagUncertainty}.
}
