% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMultinomProbs.R
\name{drawMultinomProbs}
\alias{drawMultinomProbs}
\title{drawMultinomProbs}
\usage{
drawMultinomProbs(x, nDraws = 1, nTries = 1000)
}
\arguments{
\item{x}{An object of class \code{multinom}.}

\item{nDraws}{Number of realizations of coefficient estimates to draw from \code{x}.}

\item{nTries}{Number of attempts to pull a probability matrix is all positive and has row sums equal to 1. Setting to 1 will prevent these conditions from being checked.}
}
\value{
A matrix (if \code{nDraws==1}) or list of matrices (if \code{nDraws>1}) with probabilities remote sensing patch types correspond to ground-truthing patch types. Remote sensing patch classifications are in rows and the probability classifications correspond to ground-truthed patch types are in columns.
}
\description{
A helper function to draw probabilities of classes using the distribution coefficient estimates from a multinomial model.
}
\details{
This function takes the output from \link[nnet]{multinom} and draws new values based on the estimated mean and standard error of each classification. The drawn values are then back-transformed from the logit to the probability scale to provide the probability a classification corresponds to a true class.
}
\examples{
data(substrateLayers)
# Determine which classified patch types a ground-truthing sample is most likely to overlap
withinProbs <- probWithin(joined=substrateLayers[['gtSSS']], gtSampleID="sampleID", gtPatch="gtPatch", remSensPatch="sssPatch", nDraws=100)
# Pull the highest probability matches for each sampleID
matchedSamples <- withinProbs[withinProbs[, .I[which.max(prob)], by=sampleID]$V1]
gtAccuracy <- multinom(gtPatch~sssPatch, data=matchedSamples)
drawMultinomProbs(gtAccuracy)
}
