% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPointsAndJoin.R
\name{drawPointsAndJoin}
\alias{drawPointsAndJoin}
\title{drawPointsAndJoin}
\usage{
drawPointsAndJoin(
  gtShp,
  remSensShp,
  crs = NULL,
  gtSampleID,
  gtPatch,
  remSensPatch,
  nDraws = 1000,
  gpsAccuracy = NULL,
  gpsAccuracyField = NULL,
  accuracyLevel = 0.95
)
}
\arguments{
\item{gtShp}{A point shapefile of class \code{sf} with ground-truthing data, should be in projected coordinate system (e.g., UTM, state plane).}

\item{remSensShp}{A polygon shapefile of class \code{sf} with remote sensing classifications, should be in projected coordinate system (e.g., UTM, state plane) and must match the \code{gtShp} coordinate system if \code{crs} is \code{NULL}.}

\item{crs}{Optional. CRS to be used to reproject \code{gtShp} and \code{remSensShp}.}

\item{gtSampleID}{The name of the \code{gtShp} field containing the ground-truthing sampleID.}

\item{gtPatch}{Factor, the name of the \code{gtShp} field containing the ground-truthing classifications.}

\item{remSensPatch}{Factor, the name of the \code{remSensShp} field containing the remote sensing classifications.}

\item{nDraws}{Number of random draws from an uncorrelated bivariate normal distribution used to propagate positional uncertainty.}

\item{gpsAccuracy}{Magnitude of the horizontal radius that encompasses 95\% of the data, this is also referred to as distance root mean square error. If \code{crs} is supplied, units must match the crs units, otherwise the units must match the coordinate system units of the supplied shapefiles (e.g., meters for UTM coordinate systems).}

\item{gpsAccuracyField}{The name of the \code{gtShp} field containing the ground-truthing GPS accuracy. If provided, gpsAccuracy will be ignored.}

\item{accuracyLevel}{The proportion of the true locations contained within a distance equal to \code{gpsAccuracy} from a measured GPS point.}
}
\value{
A point shapefile of class \code{sf} with ground-truthing location realizations and associated data (e.g., patch type) spatially joined to the \code{remSensShp} layer.
}
\description{
A helper function to draw ground-truthing point realizations and subset only sets of draws that have at least one ground-truthing point within each required classified patch type.
}
\examples{
data(substrateLayers)
gtShp <- substrateLayers[['gtDom_sf']]
remSensShp <- substrateLayers[['sonarClass']]
drawPointsAndJoin(gtShp=gtShp,
                  remSensShp=remSensShp,
                  gtSampleID="samplID",
                  gtPatch="domSub",
                  remSensPatch="substrate",
                  gpsAccuracy=5,
                  nDraws=100)
}
\seealso{
\link{propagateGPSUncertainty} and \link{classificationAccuracy_propagUncertainty}.
}
