% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationPointCloud.R
\name{locationPointCloud}
\alias{locationPointCloud}
\title{locationPointCloud}
\usage{
locationPointCloud(
  gtShp,
  remSensShp,
  crs = NULL,
  gtSampleID,
  gtPatch,
  remSensPatch,
  nDraws = 1000,
  nTries = 1000,
  gpsAccuracy = NULL,
  gpsAccuracyField = NULL,
  accuracyLevel = 0.95,
  subsetByProb = FALSE
)
}
\arguments{
\item{gtShp}{A point shapefile of class \code{sf} with ground-truthing data, should be in projected coordinate system (e.g., UTM, state plane).}

\item{remSensShp}{A polygon shapefile of class \code{sf} with remote sensing classifications, should be in projected coordinate system (e.g., UTM, state plane) and must match the \code{gtShp} coordinate system if \code{crs} is \code{NULL}.}

\item{crs}{Optional. CRS to be used to reproject \code{gtShp} and \code{remSensShp}.}

\item{gtSampleID}{The name of the \code{gtShp} field containing the ground-truthing sampleID.}

\item{gtPatch}{Factor, the name of the \code{gtShp} field containing the ground-truthing classifications.}

\item{remSensPatch}{Factor, the name of the \code{remSensShp} field containing the remote sensing classifications.}

\item{nDraws}{Number of random draws from an uncorrelated bivariate normal distribution used to propagate positional uncertainty.}

\item{nTries}{Maximum number of attempts used to draw a set of ground-truthing location realizations that are within all patch types classified through remote sensing.}

\item{gpsAccuracy}{Magnitude of the horizontal radius that encompasses 95\% of the data, this is also referred to as distance root mean square error. If \code{crs} is supplied, units must match the crs units, otherwise the units must match the coordinate system units of the supplied shapefiles (e.g., meters for UTM coordinate systems).}

\item{gpsAccuracyField}{The name of the \code{gtShp} field containing the ground-truthing GPS accuracy. If provided, gpsAccuracy will be ignored.}

\item{accuracyLevel}{The proportion of the true locations contained within a distance equal to \code{gpsAccuracy} from a measured GPS point.}

\item{subsetByProb}{Logical, if \code{TRUE}, each ground-truthing realization is designated as being within or outside of the corresponding remote sensing polygon using a Bernoulli trial with a probability of success equal to the probability the ground-truthing is within the patch type identified through remote sensing. Only points designated as being within a patch are retained in \code{gtRemSens_dt}.}
}
\value{
A list with: \itemize{
\item{\code{gtRemSens}} {A shapefile of class \code{sf} of spatially joined ground-truthing and remote sensing classifications.}
\item{\code{withinProbs}} {A \link[data.table]{data.table} with the probability (\code{prop}) a patch type at a ground-truth location is within a patch type classified by remote sensing for each ground-truthing sample. The number of ground-truthing location draws within a polygon (\code{N}) is also provided.}
}
}
\description{
Join ground-truthing data to patch classification from remote sensing data (e.g., side-scan sonar), calculate classification accuracy, and propagate uncertainty in classifications.
}
\details{
\code{gpsAccuracy} is passed to \link{findSD} to determine the corresponding standard deviation of the distance root
mean square error supplied to \code{gpsAccuracy}, the standard deviation of the GPS error distribution is then passed to \link{propagateGPSUncertainty}
to draw multiple realizations  of ground-truthing locations, with the number of realizations defined by \code{nDraws}. The multiple realizations of
ground-truthing points are spatially joined to the remote sensing data and supplied to \link{probWithin} to determine the probability a ground-truthing
location is within each patch type identified by remote sensing. The probability of being within  a remote sensed patch is calculated as, the proportion
of realizations within the point cloud that are within each patch identified through remote sensing. Each ground-truthing realization is then designated
as being within or outside of the corresponding remote sensing polygon using a Bernoulli process with a probability of success equal to the probability
the ground-truthing is within the patch type identified through remote sensing. Only points designated as being within a patch are retained for calculation
of a confusion matrix and classification uncertainty.
}
\examples{
data(substrateLayers)
locPointCould <- locationPointCloud(gtShp=substrateLayers[['gtDom_sf']], remSensShp=substrateLayers[['sonarClass']], crs=NULL,
                                    gtSampleID="samplID", gtPatch="domSub",
                                    remSensPatch="substrate", 
                                    nDraws=100, gpsAccuracy=5)
}
