% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchChecker.R
\name{matchChecker}
\alias{matchChecker}
\title{matchChecker}
\usage{
matchChecker(x, y)
}
\arguments{
\item{x}{A vector of values.}

\item{y}{A vector of values.}
}
\value{
A logical, \code{TRUE} if the two vectors have the same set of unique values.
}
\description{
A helper function to check if two vectors share the same unique values.
}
\examples{
dt <- data.table(V1=rep(1:10, 12), V2=sample(c("a", "b", "c"), 40, replace=TRUE))
dt[, matchChecker(x=V2, y=c("a", "b", "c")), by=V1]
}
