% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probWithin.R
\name{probWithin}
\alias{probWithin}
\title{probWithin}
\usage{
probWithin(joined, gtSampleID, gtPatch, remSensPatch, nDraws)
}
\arguments{
\item{joined}{A sf object connecting the ground-truthing data (with positional uncertainty propagated) and remote sensing classifications through a spatial join.}

\item{gtSampleID}{The name of the field containing the ground-truthing sampleID.}

\item{gtPatch}{The name of the field containing the ground-truthing classifications.}

\item{remSensPatch}{The name of the field containing the remote sensing classifications.}

\item{nDraws}{The number of draws used to propagate possible sample locations based on GPS accuracy. See \link{propagateGPSUncertainty}}
}
\value{
A \link[data.table]{data.table} with the probability (prop) a ground-truth location is within a patch type classified by remote sensing for each ground-truth sample. The number of ground-truthing location draws within a polygon (N) and patch types identified at a ground-truthing location are also provided.
}
\description{
Estimates the probability a ground-truthing data point is within a mapped polygon.
}
\details{
This is a helper function for \link{classificationAccuracy_propagUncertainty}. The probability a ground-truthing point is within a patch type identified by remote sensing is calculated as the proportion realizations for a given ground-truthing that are within each patch type identified through remote sensing that the point cloud of possible ground-truthing locations overlaps.
}
\examples{
data(substrateLayers)
withinProbs <- probWithin(joined=substrateLayers[['gtSSS']], gtSampleID="sampleID", gtPatch="gtPatch", remSensPatch="sssPatch", nDraws=100)
}
\seealso{
\link{propagateGPSUncertainty} and \link{classificationAccuracy_propagUncertainty}.
}
