% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagateGPSUncertainty.R
\name{propagateGPSUncertainty}
\alias{propagateGPSUncertainty}
\title{propagateGPSUncertainty}
\usage{
propagateGPSUncertainty(sampleID, x, y, sd, nDraws = 1000)
}
\arguments{
\item{sampleID}{A unique ID for each set of x and y coordinates.}

\item{x}{A vector of x coordinates, x coordinates units must match GPS accuracy units.}

\item{y}{A vector of y coordinates, y coordinates units must match GPS accuracy units.}

\item{sd}{Standard deviation describing GPS error distribution.}

\item{nDraws}{Number of random draws.}
}
\value{
A \link[data.table]{data.table} with X and Y coordinates for each GPS location (sampleID)
}
\description{
Draws possible locations from a bivariate normal distribution with no correlation, at given sets of x and y coordinates.
}
\examples{
x <- seq(from=10, to=100, by=10)
y <- seq(from=10, to=100, by=10)
out <- propagateGPSUncertainty(sampleID=1:10, x=x, y=y, 2.551, nDraws=100)
plot(X~Y, out)
}
