% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPatchData.R
\name{simPatchData}
\alias{simPatchData}
\title{simPatchData}
\usage{
simPatchData(
  xRange = c(0, 1000),
  yRange = c(0, 1000),
  cellsize = c(20, 20),
  crs,
  nPoints,
  patches,
  patchProbs,
  classAccu
)
}
\arguments{
\item{xRange}{Min and max x coordinates that define the horizontal boarder of a rectangular study area.}

\item{yRange}{Min and max y coordinates that define the vertical boarder of a rectangular study area.}

\item{crs}{The coordinate reference system of the simulated data.}

\item{nPoints}{Number of ground-truthing points to simulate.}

\item{patches}{A vector of character strings identifying the patch classifications to simulate.}

\item{patchProbs}{A vector that defines the proportion of the study area composed by each patch type.}

\item{cellSize}{X and Y dimensions of rectangular patches in the simulated study area.}

\item{classProb}{A square matrix that defines the probability a patch type is classified as each patch type considered. Rows contain the probability of a patch classification, given the true patch type (columns). Columns must sum to 1. Note this not the same as classification accuracy, which is the probability a classification corresponds to a true patch type.}
}
\value{
A list with: \itemize{
\item{\code{remSensPoly}} {A shapefile of class \code{sf} with simulated remote sensing polygons and classified patch types (remSensPatchType). Other fields include a unique identifier for each polygon (remSensID), the area of each polygon (gridArea), grid geometry (geometry), the patch type numeric identifier (remSensPatchID), true patch type (truePatchType), and true patch identifier (truePatchID).}
\item{\code{gtPoints}} {A shapefile of class \code{sf} with simulated ground-truthing point locations (gtPoints) and unique identifiers for each point (gtID).}
\item{\code{gtRemSensJoin}} {A shapefile of class \code{sf} with joined ground-truthing and remote sensing data. Ground-truthing patch types (gtPatchType and gtPatchID for the numeric identifier) are the simulated true patch types at the location of each ground-truthing point.}
}
}
\description{
Simulate patch classifications and ground-truthing data. Patch polygons are simulated within a rectangular grid and ground-truthing points are randomly placed within the simulated study area.
}
\examples{
simPatchData(xRange=c(0, 2500), yRange=c(0, 2500), cellsize=c(16, 16), crs="EPSG:32617", nPoints=2500,
             patches <- c("bedrock", "boulder", "cobble", "pebble", "gravel", "sand", "silt", "other"),
             patchProbs <- c(0.06, 0.04, 0.14, 0.16, 0.15, 0.25, 0.2),
             classProb <- cbind(c(0.6,  0.01, 0.05, 0.02, 0.22, 0.05, 0.05, 0.0), #bedrock
                                c(0.01, 0.48, 0.3,  0.1,  0.09, 0.01, 0.01, 0.0), #boulder
                                c(0.1,  0.2,  0.38, 0.2,  0.1,  0.01, 0.01, 0.0), #cobble
                                c(0.05, 0.05, 0.2,  0.35, 0.2,  0.1,  0.05, 0.0), #pebble
                                c(0.15, 0.05, 0.1,  0.2,  0.25, 0.2,  0.05, 0.0), #gravel
                                c(0.1,  0.05, 0.05, 0.1,  0.25, 0.35, 0.1,  0.0), #sand
                                c(0.15, 0.01, 0.01, 0.01, 0.02, 0.3,  0.5,  0.0), #silt
                                c(1/7,  1/7,  1/7,  1/7,  1/7,  1/7,  1/7,  0.0) #other
                                )
             )
}
