% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeAreas.R
\name{summarizeAreas}
\alias{summarizeAreas}
\title{summarizeAreas}
\usage{
summarizeAreas(
  propUncert_dt,
  patchType = "drawnClass",
  areaField = "area_m2",
  reClassID = "reClassID",
  sumPlot = FALSE,
  remSensShp = NULL,
  remSensPatchField = NULL,
  remSensAreaField = NULL,
  summarizeBy = NULL
)
}
\arguments{
\item{propUncert_dt}{A \link[data.table]{data.table} with reclassified patch types output from \link{classificatonAccuracy_propogateUncertainty} (will be named remSensPropUncert in the output from \link{classificatonAccuracy_propogateUncertainty}).}

\item{patchType}{The field name in \code{propUncert_dt} with the patch type of a reclassified polygon.}

\item{areaField}{The field name in \code{propUncert_dt} with the area of an individual polygon. Note that area of the polygons should be calculated and included as a field in the remote sensing polygon shapefile (the object supplied to \code{remSensShp]} \link{classificatonAccuracy_propogateUncertainty}) prior to propagating uncertainty.}

\item{reClassID}{The field name in \code{propUncert_dt} that identifies each set of patch reclassifications.}

\item{sumPlot}{Logical, should a boxplot summarizing total area of each patch type be produced.}

\item{remSensShp}{Optional. The the remote sensing polygon shapefile supplied to \code{remSensShp]} in \link{classificatonAccuracy_propogateUncertainty}. If included and \code{sumPlot=TRUE} the total area of the original patch type classifications will be included as yellow squares in the plot.}

\item{remSensPatchField}{The field name that identifies the patch type of polygons in the in shapefile supplied to \code{remSensShp}. Only required if an object is supplied to \code{remSensShp}.}

\item{remSensAreaField}{The field name that identifies the area of polygons in the in shapefile supplied to \code{remSensShp}. Only required if an object is supplied to \code{remSensShp}.}

\item{summarizeBY}{The field name that identifies the variable to summarize the area of patch types by. If \code{NULL} (the default), a summary is provided for the total study area.}
}
\value{
A \link[data.table]{data.table} summarizing the total area of patch types as the mean, median, and 0.025, 0.05, 0.25, 0.75, 0.95, and 0.975 quantiles across sets of patch reclassifications produced by \link{classificatonAccuracy_propogateUncertainty}. If \code{sumPlot=TRUE} a boxplot of summarizing the distribution of total areas of patch types across sets of patch reclassifications is also produced.
}
\description{
Summarizes the total area of reclassified patch types.
}
\details{
Note that this function assumes the area units of \code{areaField} and \code{remSensAreaField} are equal. Users must ensure area units are equal for objects supplied to \code{propUncert_dt} and \code{remSensShp}.
}
\examples{
data(substrateLayers)
# Propogate uncertainty in substrate classification
propUncert_subset <- classificationAccuracy_propagateUncertainty(gtShp=substrateLayers[['gtDom_sf']], remSensShp=substrateLayers[['sonarClass']], crs=NULL, matchMethod="bufferedSubset",
                                                                 gtSampleID="samplID", gtPatch="domSub", remSensSampleID="id",
                                                                 remSensPatch="substrate", 
                                                                 nReclass=100, gpsAccuracy=5)
summarizeAreas(propUncert_dt=propUncert_subset$remSensPropUncert, sumPlot=TRUE, remSensShp=substrateLayers[['sonarClass']], remSensPatchField="substrate", remSensAreaField="area_m2")
}
\seealso{
\link{classificationAccuracy_propagUncertainty}.
}
