% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyPoints.R
\name{surveyPoints}
\alias{surveyPoints}
\title{surveyPoints}
\usage{
surveyPoints(
  remSensShp,
  remSensPatch,
  nPatches = NULL,
  nProportion = NULL,
  nCluster = 1,
  distThreshold = NULL,
  sampleDesign = "stratified random",
  sampleAllocation = "optimum"
)
}
\arguments{
\item{remSensShp}{A polygon shapefile of class \code{sf} with remote sensing
classifications, should be in projected coordinate system (e.g., UTM, state
plane).}

\item{remSensPatch}{The name of the \code{remSensShp} field containing the
remote sensing classifications.}

\item{nPatches}{Number of ground-truthing points to survey. Omit if
\code{nProportion} is provided.}

\item{nProportion}{Proportion of classified patches to receive ground-truthing
points to survey, based on the number of patches that meet \code{distThreshold}
requirements. Omit if \code{nPatches} is provided.}

\item{nCluster}{Number of points to sample within each randomly selected patch.
If 1 (default) cluster sampling is not conducted, if greater than 1, cluster
sampling is conducted with the number of sub samples in each randomly or
stratified randomly selected polygon equal to nCluster.}

\item{distThreshold}{Minimum distance which ground-truthing survey locations
must be from a patch edge. This can be set to the accuracy of the GPS used
during remote sensing and ground-truthing surveys or some other distance to
ensure survey locations are within delineated patches. This helps reduce
co-registration errors due to positional inaccuracies, that can occur if
survey points are near patch edges.}

\item{sampleDesign}{Should survey points be placed randomly
(\code{sampleDesign="random"}) or with a stratified random design
(\code{sampleDesign="stratified random"}). Defaults to stratified random.}

\item{sampAllocation}{If a stratified random approach is used to determine
survey locations, should an optimum or proportionate allocation allocation
be used to distribute survey locations across patch types. If
\code{sampAllocation="optimum"}, a Neyman allocation is used, using the
standard deviation of patch area to minimize variance. If
\code{sampAllocation="proportionate"}, survey points are distributed across
patch types with a proportionate allocation. The Neyman allocation has the
advantage of taking number and size variation of patches into account, but
may under allocate samples to patch types with few observations or low
variability in size. Defaults to optimum sample allocation.}
}
\value{
A shapefile of class \code{sf} with x and y coordinates for a user
defined number of randomly or stratified randomly drawn survey locations.
}
\description{
Assists with designing ground-truthing surveys by drawing locations to
conduct ground-truthing data collection.
}
\details{
Survey points are placed with either a random or stratified
random sampling design and limited to areas that are a user defined
distance or more from patch edges, to reduce potential for co-registration
errors in ground-truthing surveys.
}
\examples{
data(substrateLayers)
set.seed(11252025)
gtSurvey <- surveyPoints(remSensShp = substrateLayers$sonarClass,
                         remSensPatch="substrate", sampleDesign="random", 
                         nPatches=100, distThreshold=5)

ggplot() +
  geom_sf(data=substrateLayers$sonarClass, 
          aes(fill=substrate, color=substrate)) +
  geom_sf(data=gtSurvey) +
  coord_sf(datum=st_crs(substrateLayers$sonarClass),
           ylim=c(4580800, 4581200), xlim=c(448500, 448950))
}
